chore: refactor AuditProducerRoutingKeyResolver with abstract class
